+++ lib/src/DbClientManager.h
@@ -53,7 +53,7 @@ class DbClientManager : public trantor::NonCopyable
                         const std::string &name,
                         const bool isFast,
                         const std::string &characterSet,
-                        long double timeout);
+                        double timeout);
     bool areAllDbClientsAvailable() const noexcept;
 
   private:
@@ -65,7 +65,7 @@ class DbClientManager : public trantor::NonCopyable
         ClientType dbType_;
         bool isFast_;
         size_t connectionNumber_;
-        long double timeout_;
+        double timeout_;
     };
     std::vector<DbInfo> dbInfos_;
     std::map<std::string, IOThreadStorage<orm::DbClientPtr>> dbFastClientsMap_;
