+++ lib/src/TaskTimeoutFlag.cc
@@ -15,10 +15,9 @@
 #include "TaskTimeoutFlag.h"
 using namespace drogon;
 
-TaskTimeoutFlag::TaskTimeoutFlag(
-    trantor::EventLoop *loop,
-    const std::chrono::duration<long double> &timeout,
-    std::function<void()> timeoutCallback)
+TaskTimeoutFlag::TaskTimeoutFlag(trantor::EventLoop *loop,
+                                 const std::chrono::duration<double> &timeout,
+                                 std::function<void()> timeoutCallback)
     : loop_(loop), timeout_(timeout), timeoutFunc_(timeoutCallback)
 {
 }
